Two stage-service worker
